Refactor so that MapPage is inside MapPageWrapper and MapModal added to components
