+++ frontend/src/pages/RoadTripPage/RoadTripPage.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Switch, Route, Redirect, useRouteMatch, useParams } from 'react-router-dom';
 import RoadTripTopBar from '../../components/RoadTripTopBar/RoadTripTopBar';
 import './RoadTrip.css';
-import MapPageWrapper from './MapPage/MapPageWrapper/MapPageWrapper';
+import MapPageWrapper from './MapPageWrapper/MapPageWrapper';
 import ItineraryPage from './ItineraryPage/ItineraryPage';
 import EmergencyDetailsPage from './EmergencyDetailsPage/EmergencyDetailsPage';
 import PackingListPage from './PackingListPage/PackingListPage';
